move files for future archs
