+++ fish_diffusion/archs/diffsinger/diffusions/noise_predictor.py
