--- fish_diffusion/denoisers/builder.py
@@ -1,3 +1,0 @@
-from mmengine import Registry
-
-DENOISERS = Registry("denoiser")
