+++ fish_diffusion/modules/encoders/identity.py
