+++ fish_diffusion/modules/encoders/naive_projection.py
